Multidimensional Data Filtering and Visualization
A numerical rating of sensory wine quality is given in the column“quality”. Display the distribution
of these scores with a histogram. What is the range of this score in the data? 
c) Derive a coarser classification of quality into “low”, “medium”, and “high”, by grouping together
the two lowest, the intermediate, and the two highest quality scores that occur in the dataset,
respectively. Replace the original “quality” column with a new column “quality bin” that contains
these labels. 
d) We would like to investigate differences between high and low quality wines. Therefore, create a
filtered data frame in which the medium-quality wines are omitted. 
e) Visualize all numerical attributes in a scatterplot matrix. Color the two quality levels differently.
Hint: Using seaborn, you can create this plot with a single (and simple) line of code.
f) Based on the visualization, name five attributes that appear to best distinguish between high and
low quality. 
g) Now, use an automated feature selection technique to identify five attributes that distinguish
between high and low quality. More specifically, please use the F score from a one-way analysis
of variance (ANOVA) to rank the attributes, as implemented in scikit learn’s f_classif. What
are the five best attributes according to this measure? Are they the same as those you identified
visually? Create a filtered data frame that only contains the top five attributes, plus the “quality
bin”.